+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Added task status information for each replica when a workload fails.
+
 ## [v2.6.1](https://github.com/allenai/beaker-gantry/releases/tag/v2.6.1) - 2025-06-10
 
 ### Fixed